QQmlCodeModel: avoid calling methods in wrong thread
Add some asserts to avoid calling methods from the wrong thread (this
only seems to happen in the tst_qmlls_qqmlcodemodel tests, not in qmlls
itself). Also make the methods private to avoid this from happening
again.

Adapt the tests to call the right methods, and to wait for the
QQmlCodeModel file loading thread to finish (by waiting on the newly
added openUpdateThreadFinished signal).

This fixes the flakines from tst_qmlls_qqmlcodemodel reloadLotsOfFiles,
where a data race, caused by the test calling one of QQmlCodeModels
function in the wrong thread, was making the test flaky.

Pick-to: 6.10
Fixes: QTBUG-142264
Change-Id: I6d6b89674344f271e94e6e558a34dd6900cfe703
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>