qmlls: slim down .qmlls.ini files
We only need .qmlls.ini files to contain the path to the build folder,
and the cmake-calls option (when used by the user).

Remove the outdated import path, and the documentation directory, from
the .qmlls.ini file: instead, it should be read from the
.qmlls.build.ini (which is found by qmlls with help of the .qmlls.ini
file).

Replace add_custom_command with a file(CONFIGURE), now that we don't
need to execute qtpaths anymore for the .qmlls.ini generation.

Change-Id: I2d5e7b164d68e8e092c4995916025fd04d199f12
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>