Windows style: override the application palette when in dark mode
The native Windows (XP) styling code cannot render controls in the dark
color scheme. Our code detects that the application is running in dark
mode, and then explicitly creates a light palette that gets set on the
QQuickTheme singleton.

However, the palette that the style is actually using isn't always based
on the QQuickTheme palette, so we get inconsistent colors: the window
background is dark, a button font is light, while the button itself is
also light. This makes the UI unusable.

QGuiApp has no knowledge of Qt Quick's styling infrastructure, or of
QQuickTheme, so it cannot ask the style to "polish" the palette like we
do in widgets.

As a quick fix, override the global QGuiApplication palette with the
light palette.

[ChangeLog][Controls] The Windows native style will use a light palette
as the application global palette on Windows systems running in dark
mode. The style cannot render dark controls, and mixing a dark
application palette with some UI elements rendered in light mode using
the Control style results in inconsistent and unusable user interfaces.
For Dark mode UIs, use a style that supports dark mode, like the
FluentWinUI3 or Fusion styles.

Pick-to: 6.10 6.8
Change-Id: I5e279b88ba018baa03936624e07f48e8986eb1c6
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>