qqmldomtop: take a mutex in semanticAnalysis()
It seems we have a race condition on the import paths used here, when
the import paths get updated after a build finished, while the qml file
loading thread is using the same import path. Add a mutex.

Add a semanticAnalsysUnlocked() for one use-case where the mutex is
already taken, to avoid a deadlock.

Pick-to: 6.11 6.10
Task-number: QTBUG-142352
Change-Id: I0886e2f4a61d10f7e5d869bef76ee4eca5bfa91c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>