StyleKit: only store the nested states that are actually used
It's wasteful to set off pointer space for all the possible
states that can be nested. Usually a state will not
use nested states at all. E.g the assignment
'button.pressed.spacing' would cause the 'pressed' state
to also set off space for all the states that can be nested
inside it, even if they are never used.

In addition, the current implementation would, by accident,
lazy create all the nested states while traversing up the
parent chain of states. And as it turns out, by just
storing the nested state flag inside the state, we don't
need to do any traveral at all.

This patch will make sure that we only allocate space for
nested states if we encounter that at least one of them
are being used, and store the nested state to avoid
unnecessary traversal.

Change-Id: Ied87f0c6b04057cb4c9f3e31f275a2d4b87c3efb
Reviewed-by: Doris Verria <doris.verria@qt.io>