SearchField: Prevent indicator warnings in testbench
In the SearchField testbench manual test for the macOS style, some
warnings were appearing because the x and y positions of the
indicators are used in this style, but are set to null in certain
test cases. A condition was added to check whether the indicators
are set before accessing their positions, preventing these warnings.

Fixes: QTBUG-141882
Pick-to: 6.11 6.10
Change-Id: I1c7a208ff32e570e384a2c41689cbfbb6d218be6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>