VectorImage: Add dependency on QtQuick.Effects
This amends 39ec01eec2b5262cb60f5c266d5f4107aae7e415.

On Android, we need to ensure implicit imports are known
to the tools. When adding support for the blur effect, we
also added an implicit dependency on MultiEffect, but since
no one told the deployment tool, we would get a test failure
when loading files.

Fixes: QTBUG-142472
Change-Id: Ic2f91c9fce94a90ca05314827edb77b7f9cdec4a
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>