qmllint: warn about shadowing members with ids
Implement a new warning category to warn about ids shadowing
members. Shadowing with ids is a bit tricky: ids can shadow
properties, methods and signals from unrelated QQmlJSScopes in the
same context and its inner contexts.

Add a warning that warns on ids shadowing members in the same file,
while taking care of respecting Component Root boundaries.

In addition, add a "Note:" warning that points to the definition of the
attributes that is being shadowed, if the definition is in the same file
as the id definition, or to the QmlObject that inherits it.

Disable the warnings on few tests so that they don't fail.

Fixes: QTBUG-142455
Change-Id: I5a5f3b8bb81831ce6ae45f73324707000585106c
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>