StyleKit: don't bind to the palette from the Fallback style
We used to bind several of the colors in the Fallback style
the palette. This turns out to be not such a good idea. The
reason is that the palette differs across platforms, which
means that a style created on e.g macOS can end up looking
completly different when running on Ubuntu.

While this might be wanted sometimes, more often it has the
potential to come as a surprise at a late stage during
development, and contributes to a "write once, debug everywhere"
scenario.

This patch will therefore change this. We now choose to follow
the Basic style approach, and "hard-code" the colors in the
Fallback style to be the same on all platforms. This gives
developers a stable platform to build their own styles on top
of.

It's still fully possible to bind all colors in their own
style to the palette, if they want a more dynamic style that
respects the platform palette.

Task-number: QTBUG-130067
Pick-to: 6.11
Change-Id: Ibeeedacb35b5a4c68c827b66baf0298a42b2948a
Reviewed-by: Doris Verria <doris.verria@qt.io>