Avoid create offscreen surface when using custom rhi device
When QQuickRenderControl is used for multi-threaded rendering
with a GLES backend, a warning "Attempting to create QWindow-based
QOffscreenSurface outside the gui thread. Expect failures." may be
emitted. This warning occurs when an QOffscreenSurface is created even
though the QQuickWindow already has a custom RHI (Rendering Hardware
Interface) graphics device explicitly configured.

Creating an additional offscreen surface in this specific context
is redundant and can lead to the aforementioned warning, especially
given the multi-threaded nature. This change introduces a check to
detect if the QQuickWindow is already utilizing a custom RHI graphics
device. If a custom RHI device is present, the creation of the general
QOffscreenSurface is skipped. This prevents the warning and potential
resource conflicts, ensuring QQuickRenderControl operates correctly when
an RHI-specific graphics device is already in use, without unnecessary
surface allocations.

Pick-to: 6.11
Change-Id: I343dff445ad320ab120c71764d756126bf0ce133
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>