VectorImage: Escape quotes in font families
Loading SVG files that use font families with quotes will produce
invalid output that does not parse correctly. When outputting the
font family, we have to escape the quotes first. In addition,
any character in the font family has to be valid HTML when we
output it into a style tag.

Pick-to: 6.8 6.10 6.11
Change-Id: If25b3cfd3a537d7f7c8c65045deece1ad02b43c3
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>