StyleKit: ensure we hide delegates upon setting 'visible: false'
As it stood, if you e.g did:

control.background.shadow.visible: true
control.hovered.background.shadow.visible: false

Then the shadow would stay visible also when the control
was hovered. The reason was because the shadow was already
created in the normal state, and toggling visiblity after
that point had no effect.

This patch will fix this, so that we hide the shadow if
its visibility is set to false after first having been
created. We refrain from destroying it again, since doing so
has a higher cost, and most likely, the hiding of
the shadow is just temporarily.

Task-number: QTBUG-130067
Pick-to: 6.11
Change-Id: Ib107b11b7f7fa638d7ac439348b9ccb75bf1bb1a
Reviewed-by: Doris Verria <doris.verria@qt.io>