StyleKit: only emit signals globally for readers of the correct type
When the application wrote a new value to a property in the style, we
used to emit a property changed signal for that property from every
StyleKitReader (control) in the application. We did this because we
could not easily know which controls would be affected by such a write
because of property propagation. This could therefore be very slow.

But now that we have more functionallity in StyleKit available, such
as being able to resolve a controls base types, we can now improve on
this logic.

Therefore, this patch will ensure that we instead only emit changes
from the StyleKitReaders of the same type (or a base type) as the
control written to in the style. The result is that a write from the
app, such as:

StyleKit.style.groupBox.background.radius = value

will cause an update only for groupBoxes, skipping all other controls.
This is clearly faster.

Task-number: QTBUG-130067
Pick-to: 6.11
Change-Id: I6432db93f282b8155d9270a57206f0d558b801a8
Reviewed-by: Doris Verria <doris.verria@qt.io>