Qt.createQmlObject: Trim compilation units
When we call createQmlObject, a completely new compilation unit is
created. If an object is stil in use, that compilation is obviously
needed. However, if code is using createQmlObject repeatedly with the
same URL, there's a good chance that the object was only temporarily
needed.
To avoid unbounded memory usage in that case, we call
trimCompilationUnitsForUrl, to remove the no longer needed CUs.

Note that this does not help if the URL is also changing dynamically,
but we don't want to trim all caches, including ones the user might
actually want to hold on. To handle such cases, we should rather
1. integrate the trimming logic with the gc
2. give it a separate, configurable "high water mark"
3. Use some proper caching system for CUs, e.g. LRU
That is however out of scope for this commit.

Pick-to: 6.11 6.10 6.8
Fixes: QTBUG-142555
Change-Id: I7ebb63abd9bb99531b6b6b2cf1f98b35b1e652e2
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>