Fix FileDialog with -no-accessibility build flag
ab83b2d2d34095bd56ca0e9e044374d314277321 caused the FileDialog
to stop working for -no-accessibility builds. That's because the
Accessible attached object is not recognised as an existing type
in those builds, and thus the component can't be created.

To fix it, we'll instead have to set the accessible name in C++.

As a drive-by, make all non-FINAL properties of
QQuickFileDialogImplAttached FINAL.

Fixes: QTBUG-142511
Pick-to: 6.11 6.10 6.8
Change-Id: Iea679a4acfe56e9938149dddf3410c4a55f8b28b
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>