StyleKit, optimization: track which properties a style/theme stores
This patch introduces an optimization that allows the
propagation engine to skip checking a style/theme/variation
if it does not store a value for the property being queried.

This significantly reduces the number of QHash lookups needed
during property resolution. For example, when resolving all
property values for a Button in the hovered state, the number
of “failed” lookups dropped from 1632 to 168 when tested with
the Haze style.

The large improvement is likely due to the typical separation
of concerns between styles and themes (at least in Haze):
styles primarily define structural properties such as implicit
size and corner radii, while themes define colors. Additionally,
many properties in the Fallback style remain untouched by both
style and theme. Under such conditions, which should be common,
this optimization becomes highly effective.

Task-number: QTBUG-130067
Pick-to: 6.11
Change-Id: I25d3e350a70243b8e5f96b08f9217a3404866572
Reviewed-by: Doris Verria <doris.verria@qt.io>