StyleKit: use QHash instead of QMap
It's apparently not only faster, but also more memory efficient
to use a QHash. QMap can be better if one needs to iterate
over the keys in sorted order. But that is never the case
for StyleKit. So switch to use QHash instead of QMap for
the affected containers.

Task-number: QTBUG-130067
Pick-to: 6.11
Change-Id: I6f430acbd32555cba85de8191aefa71140b07111
Reviewed-by: Doris Verria <doris.verria@qt.io>