VectorImage: Generalize defs tags as Components
<defs> tags can contain different elements that are referenced
from elsewhere in the tree. This includes filters, masks and
also arbitrary nodes that are referenced from <use> tags.

For filters and masks, we want to output these inside of
Components so that they can be referenced from Loaders later.

However, for "used" elements inside <defs>, we only want to
output these when they are used. The documentation explicitly
says to do a deep copy of the subtree and insert it at the
point of the <use> tag.

So we had hacks in place to only generate <filter> and <mask>
if we detected that we were inside a <defs>. This of course
caused problems because it would exclude any shape that might
be part of e.g. the mask.

In addition, filters and masks can also be defined outside
defs tags, and this is equivalent. This caused additional
bloat in the code.

Instead of parsing the contents of <defs> we now do a prepass
on the tree and find all elements that are referenced as
filters or masks. When markers are added later, these can
easily be added in the same way.

For each such item, we isolate a part of the node tree so
that includes its ancestors, itself and its children. Then
we traverse and output this into a Component which can then
be referenced later.

<use> tags are still implemented the same way as before, by
doing a copy, since this is specific in the W3C docs (and it
is needed to get correct style cascading etc.)

One caveat is that the filters/masks also need to generate
things outside of the Component, since we need certain
properties to be accessible from the main context. We handle
this by putting the generator in a special "defsSuffix mode".
While this is active, everything will be generated into a
string which is appended to the Component when it ends.

Fixes: QTBUG-142815
Task-number: QTBUG-121538
Change-Id: Iebcc0a8b9896c99ff98fd4ccbfc845533bb8e194
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>