QQuickItem::clipRect: Don't participate in bindings
We don't want clipRect to capture width and height. While technically
that would make sense (after all, the clipRect does change if you modify
width and height), it makes internal usage of the function problematic:
We could end up creating binding loops if it's triggered by the width or
height setter (which is the case with an upcoming change to
transformChanged).

Pick-to: 6.11
Change-Id: I3e19a9687a025e35969d16f300ca2d1b6fc8a624
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>