macOS: adjust the button margins to ensure the title is centered
The title (content) margins for a button had different margins
at the top compared to the bottom. The result was that the
label didn't end up centered on the button on macOS 26.

This patch will adjust the button title margins so that the
button label ends up centered on macOS 26. It will also
slighly adjust the margins on macOS16.4, and earlier, to make
label (more) centered on those platforms as well.

Pick-to: 6.11 6.10 6.8 6.5
Change-Id: Icd7571202add65ec312787b0aa94fd1bc49ef72a
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>