Qml override semantics: fix OverridingNonVirtual
For some reason the logic of overrides when virtual keyword is missing
was incorrect: after some discussions during the review, misleading
FinalOverridingNonVirtual was introduced.
In fact what was required to do is to differentiate between the "error"
case: Derived {override ... p} Base {... p} and "warning" case
Derived {(final) ... p} Base {... p}

This patch aims to fix the issue in line with the intended and documented
semantics.

Task-number: QTBUG-98320
Pick-to: 6.11
Change-Id: Ia738fc57f80d8d3a299f052f8bd6399b9ad8b428
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>