svgtoqml: Fix running from packaged build
In a packaged build, svgtoqml is placed next to a qt.conf file.
This caused it to fail to launch from anywhere except in its local
directory, with an error that it could not find the QPA plugin.

It turns out that this was triggered by trying to query the
available QPA plugins before creating a QGuiApplication. We did
this to avoid forcing the minimal plugin if it is not available,
but other tools that hardcode this plugin does it unconditionally,
so we do the same in svgtoqml.

Pick-to: 6.11
Task-number: QTBUG-142792
Change-Id: I2b7c0aa20c251a8fbf4f2372b8bc5e0802e298b2
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>