VectorImage: Add environment var to freeze animations for testing
Baseline testing requires animated scenes to stabilize in a consistent
state so that a deterministic screenshot can be taken. That is not
always feasible for animated vector images. To facilitate such
testing, introduce an environment variable that forces VectorImage to
start up in paused state at a specific time point in the animation.

Pick-to: 6.11
Change-Id: I0d66bb5ae2550de9c81f6a32b451c872fe5f218f
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>