QQuickCanvasItem: fix QUrl::resolved() usage
The result of QUrl::fromLocalFile() is a url with `file:` scheme, so
the baseUrl is ignored.

Fix it, as Thiago pointed out, by using QUrl::setPath() to accommodate
file names that contain special characters, such as `#` (we don't want
`#` to be interpreted as a fragment).

The unittests worked because the `filename` arg it passed to
QQCItem::save() is always either absolute or relative to the current
working dir.

Amends fdc87143eb9eff98938b6a086c3e81432be08e4d.

Pick-to: 6.11 6.10 6.8 6.5
Change-Id: Id23f03f46651ace5cb3a3e2c4271e4c2bfc61b4f
Reviewed-by: David Faure <david.faure@kdab.com>